@@ -14,15 +14,20 @@
         }
     ],
     "require": {
-        "php": "^5.3.3 || ^7.0 || ^8.0",
+        "php": "^7.2 || ^8.0",
         "symfony/polyfill-ctype": "^1.8"
     },
-    "require-dev": {
-        "phpunit/phpunit": "^4.8.36 || ^7.5.13"
-    },
     "conflict": {
-        "vimeo/psalm": "<3.9.1",
-        "phpstan/phpstan": "<0.12.20"
+        "phpstan/phpstan": "<0.12.20",
+        "vimeo/psalm": "<4.6.1 || 4.6.2"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "^8.5.13"
+    },
+    "extra": {
+        "branch-alias": {
+            "dev-master": "1.10-dev"
+        }
     },
     "autoload": {
         "psr-4": {
